Added DataUtility
Commented out conection.cs and added Data Utility
