+++ Data/Connection.cs
@@ -5,33 +5,33 @@ using System.Threading.Tasks;
 using Microsoft.Extensions.Configuration;
 using Npgsql;
 
-namespace RockwellBlog.Data
-{
-    public class Connection
-    {
-        public static string GetConnectionString(IConfiguration configuration)
-        {
-            //var connectionString = configuration.GetConnectionString("DefaultConnection");
-            var databaseUrl = Environment.GetEnvironmentVariable("DATABASE_URL");
-            return string.IsNullOrEmpty(databaseUrl) ? configuration.GetConnectionString("DefaultConnection") : BuildConnectionString(databaseUrl);
-        }
+//namespace RockwellBlog.Data
+//{
+//    public class Connection
+//    {
+//        public static string GetConnectionString(IConfiguration configuration)
+//        {
+//            var connectionString = configuration.GetConnectionString("DefaultConnection");
+//            var databaseUrl = Environment.GetEnvironmentVariable("DATABASE_URL");
+//            return string.IsNullOrEmpty(databaseUrl) ? configuration.GetConnectionString("DefaultConnection") : BuildConnectionString(databaseUrl);
+//        }
 
-        private static string BuildConnectionString(string databaseUrl) 
-        {
-            var databaseUri = new Uri(databaseUrl);
-            var userInfo = databaseUri.UserInfo.Split(":");
+//        private static string BuildConnectionString(string databaseUrl) 
+//        {
+//            var databaseUri = new Uri(databaseUrl);
+//            var userInfo = databaseUri.UserInfo.Split(":");
 
-            return new NpgsqlConnectionStringBuilder()
-            {
-                Host = databaseUri.Host,
-                Port = databaseUri.Port,
-                Username = userInfo[0],
-                Password = userInfo[1],
-                Database = databaseUri.LocalPath.TrimStart('/'),
-                SslMode = SslMode.Prefer,
-                TrustServerCertificate = true
-            }.ToString();
-        }
+//            return new NpgsqlConnectionStringBuilder()
+//            {
+//                Host = databaseUri.Host,
+//                Port = databaseUri.Port,
+//                Username = userInfo[0],
+//                Password = userInfo[1],
+//                Database = databaseUri.LocalPath.TrimStart('/'),
+//                SslMode = SslMode.Prefer,
+//                TrustServerCertificate = true
+//            }.ToString();
+//        }
 
-    }
-}
+//    }
+//}
